delete btn is fumctional now.

+++ src/pages/MyCart.jsx
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react";
-import toast, { Toaster } from "react-hot-toast";
+import toast from "react-hot-toast";
+import swal from 'sweetalert';
+
 
 
 const MyCart = () => {
@@ -9,33 +11,64 @@ const MyCart = () => {
             .then(resp => resp.json())
             .then(data => setProducts(data))
     }, []);
-    const handleDelete = id => {
-        fetch(`http://localhost:5173/myCart/${id}`,{
-            method: 'DELETE'
+
+    const handleDelete = _id => {
+        // console.log(_id);
+
+        //sweetheart for confirmation
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
         })
-        .then(resp=>resp.json())
-        .then(data=>{
-            if(data.deletedCount) toast.error("Deleted."); 
-        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    //delete for db by server
+                    fetch(`http://localhost:5000/cart/${_id}`, {
+                        method: 'DELETE'
+                    })
+                        .then(resp => resp.json())
+                        .then(data => {
+                            if (data.deletedCount){
+                                toast.error("Deleted.");
+                            }
+                        });
+                        const remaining = products.filter(product => product._id !==_id);
+                        setProducts(remaining);
+                }
+                else {
+                    toast("Canceled.")
+                }
+            });
+
     }
     return (
         <div>
             <h1 className="text-accent text-4xl mb-10 text-center">Your Products:</h1>
             {/* cards */}
             <div className="flex flex-col gap-3">{
-                products.map(product =>
+                products.map( product =>
+                    //card start here
                     <div className="card w-96 bg-base-100 shadow-xl image-full mx-auto" key={product._id}>
-                        <figure><img src={product.img} alt={product.name} /></figure>
+                        {/* card bg */}
+                        <figure>
+                            <img src={product.img} alt={product.name} />
+                        </figure>
+                        {/* texts */}
                         <div className="card-body">
                             <h2 className="card-title text-primary text-3xl">{product.name}</h2>
-                            <p className="text-xl">{product.product_type}<span className="badge badge-accent ml-5">${product.price}</span></p>
+
+                            <p className="text-xl">
+                                {product.product_type}
+                                <span className="badge badge-accent ml-5">${product.price}</span>
+                            </p>
+                            {/* btn */}
                             <div className="card-actions justify-end">
-                                <button onClick={()=> handleDelete(product._id)} className="btn btn-secondary">Delete</button>
+                                <button onClick={() => handleDelete(product._id)} className="btn btn-secondary">Delete</button>
                             </div>
                         </div>
-                        <div><Toaster
-                            position="top-right"
-                            reverseOrder={false}/></div>
                     </div>)
             }</div>
         </div>
